% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aranet4r-package.R
\docType{package}
\name{aranet4r-package}
\alias{aranet4r}
\alias{aranet4r-package}
\title{aranet4r: Import, combine and manipulate Aranet4 Carbon Dioxide measurements}
\description{
Use this package to combine CSVs that have been exported from your Aranet4 device and make graphs of the results.
}
\author{
\strong{Maintainer}: Sam Herniman \email{herniman.sam@gmail.com} (\href{https://orcid.org/0000-0001-7663-2897}{ORCID})

}
\keyword{internal}
