% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aranet.R
\name{read_aranet}
\alias{read_aranet}
\title{Read and combine CSVs exported from Aranet4}
\usage{
read_aranet(x)
}
\arguments{
\item{x}{A path to a directory or a list of paths to CSVs}
}
\value{
A data.table data.frame containing Aranet4 data
}
\description{
Read and combine CSVs exported from Aranet4
}
\details{
If you give this function a single CSV it will
read it in standardize all the columns and add some metadata
from the file name. If you give this function a directory
containing multiple CSVs or a vector of paths to multiple CSVs,
it will read and standardize all the supplied files into a single
dataframe. It will also remove duplicate rows allowing one to export
results from their device multiple times and then combine them
into a single file.
}
\examples{
\dontrun{
# Read every CSV in a directory
dt <- read_aranet("path/to/directory/of/csvs")
# Read only specified CSVs
x <- c(
  "path/to/aranet_file_one.csv",
  "path/to/aranet_file_two.csv"
)
dt <- read_aranet(x)
}
}
