% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aranet.R
\name{read_one_aranet}
\alias{read_one_aranet}
\title{Read CSVs from Aranet4}
\usage{
read_one_aranet(x)
}
\arguments{
\item{x}{A path to a single CSV file}
}
\value{
A data.table data.frame
}
\description{
Read CSVs from Aranet4
}
